Added working householder

+++ BackwardSubstitutionSolver.cs
@@ -8,7 +8,7 @@ namespace MatrixSolver;
 
 public class BackwardSubstitutionSolver : ISolver
 {
-    public decimal[] Solve(IMatrix matrix, decimal[] parameters)
+    public double[] Solve(IMatrix matrix, double[] parameters)
     {
         AssertUpperTriangularSystem(matrix);
         if (matrix.Rows != parameters.Length)
@@ -16,7 +16,7 @@ public class BackwardSubstitutionSolver : ISolver
             throw new ArgumentException($"{nameof(parameters)} does not match the matrix size.");
         }
 
-        var solvedValues = new decimal[matrix.Columns];
+        var solvedValues = new double[matrix.Columns];
         for (var row = matrix.Rows - 1; row >= 0; row--)
         {
             var rowSum = parameters[row];
@@ -36,7 +36,7 @@ public class BackwardSubstitutionSolver : ISolver
         {
             for (var column = 0; column < row; column++)
             {
-                if (matrix[column, row] != 0M)
+                if (matrix[column, row] != 0D)
                 {
                     throw new ArgumentException($"{nameof(BackwardSubstitutionSolver)} expects a upper triangular system!");
                 }
